web app: refactor version handling to use in all BPs
Version handling is now part of GeneralView from where
all the views inherit
